+++ CHANGELOG.md
@@ -4,6 +4,9 @@
     * Add `Tm1637Display` class to support LED modules using the TM1637 chip.
     * Simplify `LedDisplay` base class API.
     * Add `TemperatureWriter`.
+    * Add `remapArray` to `Tm1637Display::begin()` to support logical reordering
+      of the LED digits to support certain 6-digit LED modules which are
+      physically wired out of order.
 * 0.4 (2021-04-09)
     * A complete refactoring of the previous v0.3 version, which I could not
       understand anymore.